pyproject.toml
[project]
name = "chatbot_ia_assistent"
version = "0.1.0"
description = "A Telegram bot with AI and database integration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.32.5"
    , "python-dotenv>=1.1.1"
    , "openai>=2.0.0"
    , "python-telegram-bot>=22.5"
    , "psycopg2-binary>=2.9.10"
    , "langchain>=0.3.0"
    , "langchain-openai>=0.3.0"
    , "langchain-community>=0.3.0"
    , "pydantic>=2.11.9"
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.2"
    , "pytest-asyncio>=0.24.0"
    ,
]
dev = []

[tool.setuptools]
packages = ["src.bot", "src.config", "src.database", "src.schemas", "src.services", "src.utils"]

[tool.setuptools.package-dir]
"src.bot" = "src/bot"
"src.config" = "src/config"
"src.database" = "src/database"
"src.schemas" = "src/schemas"
"src.services" =  "src/services"
"src.utils" = "src/utils"